gpio: dw: Change type of base_addr to uintptr_t
The address variable is declared in 32-bits, however, the type declaration
is changed to allow the address width of the architecture.

+++ drivers/gpio/gpio_dw.c
@@ -35,18 +35,18 @@ static int gpio_dw_port_clear_bits_raw(const struct device *port,
  * Other architectures as ARM and x86 configure IP through MMIO registers
  */
 #ifdef GPIO_DW_IO_ACCESS
-static inline uint32_t dw_read(uint32_t base_addr, uint32_t offset)
+static inline uint32_t dw_read(uintptr_t base_addr, uint32_t offset)
 {
 	return sys_in32(base_addr + offset);
 }
 
-static inline void dw_write(uint32_t base_addr, uint32_t offset,
+static inline void dw_write(uintptr_t base_addr, uint32_t offset,
 			    uint32_t val)
 {
 	sys_out32(val, base_addr + offset);
 }
 
-static void dw_set_bit(uint32_t base_addr, uint32_t offset,
+static void dw_set_bit(uintptr_t base_addr, uint32_t offset,
 		       uint32_t bit, bool value)
 {
 	if (!value) {
@@ -56,18 +56,18 @@ static void dw_set_bit(uint32_t base_addr, uint32_t offset,
 	}
 }
 #else
-static inline uint32_t dw_read(uint32_t base_addr, uint32_t offset)
+static inline uint32_t dw_read(uintptr_t base_addr, uint32_t offset)
 {
 	return sys_read32(base_addr + offset);
 }
 
-static inline void dw_write(uint32_t base_addr, uint32_t offset,
+static inline void dw_write(uintptr_t base_addr, uint32_t offset,
 			    uint32_t val)
 {
 	sys_write32(val, base_addr + offset);
 }
 
-static void dw_set_bit(uint32_t base_addr, uint32_t offset,
+static void dw_set_bit(uintptr_t base_addr, uint32_t offset,
 		       uint32_t bit, bool value)
 {
 	if (!value) {
@@ -78,12 +78,12 @@ static void dw_set_bit(uint32_t base_addr, uint32_t offset,
 }
 #endif
 
-static inline int dw_base_to_block_base(uint32_t base_addr)
+static inline int dw_base_to_block_base(uintptr_t base_addr)
 {
 	return (base_addr & 0xFFFFFFC0);
 }
 
-static inline int dw_derive_port_from_base(uint32_t base_addr)
+static inline int dw_derive_port_from_base(uintptr_t base_addr)
 {
 	uint32_t port = (base_addr & 0x3f) / 12U;
 	return port;
@@ -94,7 +94,7 @@ static inline int dw_interrupt_support(const struct gpio_dw_config *config)
 	return ((int)(config->irq_num) > 0U);
 }
 
-static inline uint32_t dw_get_ext_port(uint32_t base_addr)
+static inline uint32_t dw_get_ext_port(uintptr_t base_addr)
 {
 	uint32_t ext_port;
 
@@ -118,7 +118,7 @@ static inline uint32_t dw_get_ext_port(uint32_t base_addr)
 	return ext_port;
 }
 
-static inline uint32_t dw_get_data_port(uint32_t base_addr)
+static inline uint32_t dw_get_data_port(uintptr_t base_addr)
 {
 	uint32_t dr_port;
 
@@ -142,7 +142,7 @@ static inline uint32_t dw_get_data_port(uint32_t base_addr)
 	return dr_port;
 }
 
-static inline uint32_t dw_get_dir_port(uint32_t base_addr)
+static inline uint32_t dw_get_dir_port(uintptr_t base_addr)
 {
 	uint32_t ddr_port;
 
@@ -172,8 +172,8 @@ static int gpio_dw_pin_interrupt_configure(const struct device *port,
 					   enum gpio_int_trig trig)
 {
 	const struct gpio_dw_config *config = port->config;
-	uint32_t base_addr = dw_base_to_block_base(config->base_addr);
-	uint32_t port_base_addr = config->base_addr;
+	uintptr_t base_addr = dw_base_to_block_base(config->base_addr);
+	uintptr_t port_base_addr = config->base_addr;
 	uint32_t dir_port = dw_get_dir_port(port_base_addr);
 	uint32_t data_port = dw_get_data_port(port_base_addr);
 	uint32_t dir_reg;
@@ -235,7 +235,7 @@ static inline void dw_pin_config(const struct device *port,
 				 uint32_t pin, int flags)
 {
 	const struct gpio_dw_config *config = port->config;
-	uint32_t base_addr = dw_base_to_block_base(config->base_addr);
+	uintptr_t base_addr = dw_base_to_block_base(config->base_addr);
 	uint32_t port_base_addr = config->base_addr;
 	uint32_t dir_port = dw_get_dir_port(port_base_addr);
 	bool pin_is_output, need_debounce;
@@ -266,7 +266,7 @@ static inline void dw_pin_config(const struct device *port,
 static void gpio_dw_set_hw_mode(const struct device *port, gpio_pin_t pin, bool hw_mode)
 {
 	const struct gpio_dw_config *config = port->config;
-	uint32_t base_addr = dw_base_to_block_base(config->base_addr);
+	uintptr_t base_addr = dw_base_to_block_base(config->base_addr);
 	uint32_t port_id = dw_derive_port_from_base(config->base_addr);
 	uint32_t ctl_port;
 
@@ -336,8 +336,8 @@ static inline int gpio_dw_config(const struct device *port,
 static int gpio_dw_port_get_raw(const struct device *port, uint32_t *value)
 {
 	const struct gpio_dw_config *config = port->config;
-	uint32_t base_addr = dw_base_to_block_base(config->base_addr);
-	uint32_t port_base_addr = config->base_addr;
+	uintptr_t base_addr = dw_base_to_block_base(config->base_addr);
+	uintptr_t port_base_addr = config->base_addr;
 	uint32_t ext_port = dw_get_ext_port(port_base_addr);
 
 	*value = dw_read(base_addr, ext_port);
@@ -349,8 +349,8 @@ static int gpio_dw_port_set_masked_raw(const struct device *port,
 				       uint32_t mask, uint32_t value)
 {
 	const struct gpio_dw_config *config = port->config;
-	uint32_t base_addr = dw_base_to_block_base(config->base_addr);
-	uint32_t port_base_addr = config->base_addr;
+	uintptr_t base_addr = dw_base_to_block_base(config->base_addr);
+	uintptr_t port_base_addr = config->base_addr;
 	uint32_t data_port = dw_get_data_port(port_base_addr);
 	uint32_t pins;
 
@@ -364,8 +364,8 @@ static int gpio_dw_port_set_masked_raw(const struct device *port,
 static int gpio_dw_port_set_bits_raw(const struct device *port, uint32_t mask)
 {
 	const struct gpio_dw_config *config = port->config;
-	uint32_t base_addr = dw_base_to_block_base(config->base_addr);
-	uint32_t port_base_addr = config->base_addr;
+	uintptr_t base_addr = dw_base_to_block_base(config->base_addr);
+	uintptr_t port_base_addr = config->base_addr;
 	uint32_t data_port = dw_get_data_port(port_base_addr);
 	uint32_t pins;
 
@@ -380,8 +380,8 @@ static int gpio_dw_port_clear_bits_raw(const struct device *port,
 				       uint32_t mask)
 {
 	const struct gpio_dw_config *config = port->config;
-	uint32_t base_addr = dw_base_to_block_base(config->base_addr);
-	uint32_t port_base_addr = config->base_addr;
+	uintptr_t base_addr = dw_base_to_block_base(config->base_addr);
+	uintptr_t port_base_addr = config->base_addr;
 	uint32_t data_port = dw_get_data_port(port_base_addr);
 	uint32_t pins;
 
@@ -395,8 +395,8 @@ static int gpio_dw_port_clear_bits_raw(const struct device *port,
 static int gpio_dw_port_toggle_bits(const struct device *port, uint32_t mask)
 {
 	const struct gpio_dw_config *config = port->config;
-	uint32_t base_addr = dw_base_to_block_base(config->base_addr);
-	uint32_t port_base_addr = config->base_addr;
+	uintptr_t base_addr = dw_base_to_block_base(config->base_addr);
+	uintptr_t port_base_addr = config->base_addr;
 	uint32_t data_port = dw_get_data_port(port_base_addr);
 	uint32_t pins;
 
@@ -421,7 +421,7 @@ static void gpio_dw_isr(const struct device *port)
 {
 	struct gpio_dw_runtime *context = port->data;
 	const struct gpio_dw_config *config = port->config;
-	uint32_t base_addr = dw_base_to_block_base(config->base_addr);
+	uintptr_t base_addr = dw_base_to_block_base(config->base_addr);
 	uint32_t int_status;
 
 	int_status = dw_read(base_addr, INTSTATUS);
@@ -446,7 +446,7 @@ static DEVICE_API(gpio, api_funcs) = {
 static int gpio_dw_initialize(const struct device *port)
 {
 	const struct gpio_dw_config *config = port->config;
-	uint32_t base_addr;
+	uintptr_t base_addr;
 
 	if (dw_interrupt_support(config)) {
